requirements.txt
argcomplete==3.5.3
asgiref==3.8.1
cfgv==3.4.0
charset-normalizer==3.4.1
colorama==0.4.6
commitizen==4.4.1
decli==0.6.2
distlib==0.3.9
Django==5.1.7
django-appconf==1.1.0
django-compressor==4.5.1
django-libsass==0.9
django-sass-processor==1.4.1
django-stubs==5.1.3
django-stubs-ext==5.1.3
filelock==3.18.0
identify==2.6.9
Jinja2==3.1.6
libsass==0.23.0
MarkupSafe==3.0.2
nodeenv==1.9.1
packaging==24.2
pillow==10.4.0
platformdirs==4.3.7
pre_commit==4.2.0
prompt_toolkit==3.0.50
pytz==2025.1
PyYAML==6.0.2
questionary==2.1.0
rcssmin==1.1.2
rjsmin==1.2.2
sqlparse==0.5.3
termcolor==2.5.0
tomlkit==0.13.2
types-PyYAML==6.0.12.20241230
typing_extensions==4.12.2
tzdata==2025.2
virtualenv==20.29.3
wcwidth==0.2.13
django-simple-captcha==0.6.2
django-email-verification==0.1.0
django-otp==1.3.0
